+++ src/functions/unitary/trig/inverse/Asec.java
@@ -1,7 +1,9 @@
 package functions.unitary.trig.inverse;
 
+import config.Settings;
 import functions.GeneralFunction;
 import functions.binary.Pow;
+import functions.binary.integer.Modulo;
 import functions.commutative.Sum;
 import functions.commutative.Product;
 import functions.unitary.piecewise.Abs;
@@ -52,5 +54,17 @@ public class Asec extends InverseTrigFunction {
 
 	public Class<?> getInverse() {
 		return Sec.class;
+	}
+
+	@Override
+	public GeneralFunction simplifyInverse() {
+		if (operand.getClass().isAssignableFrom(getInverse())) {
+			GeneralFunction insideFunction = ((UnitaryFunction) operand).operand;
+			if (Settings.enforceDomainAndRange)
+				return DefaultFunctions.subtract(DefaultFunctions.PI, new Abs(DefaultFunctions.subtract(new Modulo(DefaultFunctions.TWO_PI, insideFunction), DefaultFunctions.PI)));
+			else
+				return insideFunction;
+		} else
+			return this;
 	}
 }